Add all interfaces, within formulation and start to debuging

+++ GenerateTestOfNumbers/answer.cpp
@@ -3,33 +3,34 @@
 ***********************************************************/
 
 #include "answer.h"
+#include <QString>
 #include <iostream>
 
-const int MAX_LENGHT_OF_ANSWER = 50;
+const qint16 MAX_LENGHT_OF_ANSWER = 50;
 
 //Конструктор класса со значениями по умолчанию
-Answer::Answer( int inputAnswer, int inputRadix ){
+Answer::Answer( qint16 inputAnswer, qint16 inputRadix ){
     setAnswer( inputAnswer );
     setRadix( inputRadix );
 }
 
 //Передаем значение числа в класс Answer
-void Answer::setAnswer( int inputAnswer ){
+void Answer::setAnswer( qint16 inputAnswer ){
     answer = ( inputAnswer > 0 && inputAnswer < 4097 ) ? inputAnswer : 0;
 }
 
 //Получаем значение числа из класса Answer
-int Answer::getAnswer() const {
+qint16 Answer::getAnswer() const {
     return answer;
 }
 
 //Передаем значение основания системы счисления в класс Answer
-void Answer::setRadix( int inputRadix ){
+void Answer::setRadix( qint16 inputRadix ){
     radix = ( inputRadix > 0 && inputRadix < 33 ) ? inputRadix : 0;
 }
 
 //Получаем значение основания системы счисления из класса Answer
-int Answer::getRadix() const {
+qint16 Answer::getRadix() const {
     return radix;
 }
 
@@ -37,9 +38,9 @@ int Answer::getRadix() const {
 char* Answer::translateToRadix() const {
     char *numberPtr = new char[MAX_LENGHT_OF_ANSWER];
     const char *DIGITS = "0123456789ABCDEFGHIJKLMNOPQRSTUVWXYZ";
-    int number = answer;
-    int outputRadix = radix;
-    int count, lenght = 0;
+    qint16 number = answer;
+    qint16 outputRadix = radix;
+    qint16 count, lenght = 0;
     for ( count = 0; number >= outputRadix; count++ ){
         numberPtr[count] = DIGITS[number % outputRadix];
         number /= outputRadix;
