+++ GenerateTestOfNumbers/answer.h
@@ -4,20 +4,21 @@
 #ifndef ANSWER_H
 #define ANSWER_H
 
+#include <QString>
 #include <iostream>
 
 class Answer {
     friend std::ostream &operator<<( std::ostream&, const Answer&);
     public:
-        Answer( int = 0, int = 0 );
-        void setAnswer( int );
-        void setRadix( int );
-        int getAnswer() const;
-        int getRadix() const;
+        Answer( qint16 = 0, qint16 = 0 );
+        void setAnswer( qint16 );
+        void setRadix( qint16 );
+        qint16 getAnswer() const;
+        qint16 getRadix() const;
     private:
         char* translateToRadix() const;
-        int answer;
-        int radix;
+        qint16 answer;
+        qint16 radix;
 };
 
 #endif // ANSWER_H
